Remove deprecated component and clean some code
