--- apps/app/src/plans/plans/show-plan/filters/index.tsx
@@ -1,13 +1,0 @@
-import { FiltersMenuButton as GenericFiltersMenuButton } from '@/app/ui/lists/DEPRECATED_filter-badges';
-import { Menu } from './Menu';
-import { usePlanFilters } from './plan-filters.context';
-
-export const FiltersMenuButton = () => {
-  const { filtersCount } = usePlanFilters();
-
-  return (
-    <GenericFiltersMenuButton activeFiltersCount={filtersCount} size="sm">
-      <Menu />
-    </GenericFiltersMenuButton>
-  );
-};
