+++ apps/app/src/ui/lists/DEPRECATED_filter-badges/index.ts
@@ -1,4 +1,3 @@
 export { default } from './DEPRECATED_filter-badges';
-export { FiltersMenuButton } from './FiltersMenuButton';
 export { useFiltersToBadges } from './use-filters-to-badges';
 export type { BadgeFilters, CustomFilterBadges } from './use-filters-to-badges';
